# internal/handler/handler.go
package handler

import (
	"encoding/json"
	"errors"
	"net/http"
	"strconv"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/cors"
	"github.com/google/uuid"
	httpSwagger "github.com/swaggo/http-swagger/v2"

	"wordsGo_v2/internal/middleware"
	"wordsGo_v2/internal/models"
	"wordsGo_v2/internal/service"
	"wordsGo_v2/internal/utils"
	"wordsGo_v2/slogger"
)

const ctxWithTimeout time.Duration = time.Second * 5

type Handler struct {
	wordsService service.WordsServiceI
	userService  service.UserService
}

func NewHandler(ws service.WordsServiceI, us service.UserService) *Handler {
	return &Handler{wordsService: ws, userService: us}
}

func RegisterRoutes(h *Handler, jwtSecret string) *chi.Mux {
	r := chi.NewRouter()

	r.Use(cors.Handler(cors.Options{
		AllowedOrigins:   []string{"https://*", "http://*"}, // Разрешаем откуда угодно (для разработки)
		AllowedMethods:   []string{"GET", "POST", "PUT", "DELETE", "OPTIONS"},
		AllowedHeaders:   []string{"Accept", "Authorization", "Content-Type", "X-CSRF-Token"},
		ExposedHeaders:   []string{"Link"},
		AllowCredentials: true,
		MaxAge:           300,
	}))

	r.Use(middleware.LoggerMiddleware)
	r.Use(middleware.TimeoutMiddleware(ctxWithTimeout))

	r.Get("/swagger/*", httpSwagger.WrapHandler)

	r.Route("/api/v1/", func(r chi.Router) {

		r.Route("/users", func(r chi.Router) {
			r.Post("/create", h.Create)
			r.Post("/login", h.Login)

			r.Group(func(r chi.Router) {
				r.Use(middleware.AuthMidleware(jwtSecret))

				//r.Get("/{id}", h.GetUserByID)
				//r.Get("/all", h.GetUsers)

				r.Put("/{id}", h.Update)
				r.Delete("/{id}", h.Delete)

			})
		})
		r.Group(func(r chi.Router) {
			r.Use(middleware.AuthMidleware(jwtSecret))

			r.Route("/words", func(r chi.Router) {
				r.Post("/create", h.NewWord)
				r.Post("/translate", h.Translate)
			})

			r.Route("/lesson", func(r chi.Router) {
				r.Get("/start", h.StartLesson)
				r.Post("/check", h.CheckAnswer)
				r.Post("/finish", h.Finish)
			})
		})

	})

	return r
}

// Login
// @Summary User Login
// @Description Login and get JWT token
// @Tags users
// @Accept json
// @Produce json
// @Param input body models.LoginRequest true "Login credentials"
// @Success 200 {object} models.LoginResponse
// @Failure 401 {object} handler.JSONError "Invalid credentials"
// @Router /login [post]
func (h *Handler) Login(w http.ResponseWriter, r *http.Request) {
	var req models.LoginRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		WriteError(w, http.StatusBadRequest, "Invalid request body")
		return
	}

	ctx := r.Context()
	slogger.Log.DebugContext(ctx, "Login request", "req", req)

	token, err := h.userService.Login(ctx, req)
	slogger.Log.DebugContext(ctx, "Login response", "token", token)
	if err != nil {
		if errors.Is(err, models.ErrInvalidCredentials) {
			WriteError(w, http.StatusUnauthorized, "Invalid email or password")
			return
		}
		WriteError(w, http.StatusInternalServerError, "Internal server error")
		slogger.Log.ErrorContext(ctx, "Login failed", "err", err)
		return
	}

	JSONResponse(w, http.StatusOK, models.LoginResponse{Token: token})
}

// Create creates a new user
// @Summary Create a new user
// @Description Register a new user in the system
// @Tags users
// @Accept json
// @Produce json
// @Param input body models.CreateUserRequest true "User registration info"
// @Success 201 {object} models.UserResponse
// @Failure 400 {object} handler.JSONError "Invalid input or validation info"
// @Failure 409 {object} handler.JSONError "User already exists"
// @Failure 500 {object} handler.JSONError "Internal server error"
// @Router /users [post]
func (h *Handler) Create(w http.ResponseWriter, r *http.Request) {

	var req models.CreateUserRequest

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		WriteError(w, http.StatusBadRequest, "Invalid request body")
		slogger.Log.ErrorContext(r.Context(), "Invalid request body", "err", err)
		return
	}

	if req.Email == "" || req.Password == "" || req.FirstName == "" || req.LastName == "" {
		WriteError(w, http.StatusBadRequest, "Fields cannot be empty")
		return
	}
	if err := utils.ValidatePassword(req.Password); err != nil {
		WriteError(w, http.StatusBadRequest, err.Error())
		return
	}

	ctx := r.Context()

	slogger.Log.DebugContext(ctx, "Creating user",
		"email", req.Email,
		"firstName", req.FirstName,
		"lastName", req.LastName,
	)

	createdUser, err := h.userService.Create(ctx, req)
	if err != nil {
		switch {
		case errors.Is(err, models.ErrUserAlreadyExists):
			WriteError(w, http.StatusConflict, "User already exists")
			return
		}
		WriteError(w, http.StatusInternalServerError, err.Error())
		slogger.Log.ErrorContext(ctx, "Failed to create user", "err", err)
		return
	}

	JSONResponse(w, http.StatusCreated, createdUser)
}

// GetUserByID Get User By ID
// @Summary Get user by ID
// @Description Get User By ID from DB
// @Tags users
// @Accept json
// @Produce json
// @Param id path string true "User UUID" format(uuid)
// @Success 200 {object} models.UserResponse
// @Failure 400 {object} handler.JSONError "Invalid UUID"
// @Failure 404 {object} handler.JSONError "User not found"
// @Failure 500 {object} handler.JSONError "Internal server error"
// @Router /users/{id} [get]
func (h *Handler) GetUserByID(w http.ResponseWriter, r *http.Request) {

	id, err := uuid.Parse(chi.URLParam(r, "id"))
	if err != nil {
		WriteError(w, http.StatusBadRequest, "Invalid user ID")
		return
	}

	ctx := r.Context()

	user, err := h.userService.GetUserByID(ctx, id)
	if err != nil {
		switch {
		case errors.Is(err, models.ErrUserNotFound):
			WriteError(w, http.StatusNotFound, "User not found")
			slogger.Log.DebugContext(ctx, "User not found", "err", err, "id", id)
			return
		default:
			WriteError(w, http.StatusInternalServerError, "Failed to get user")
			slogger.Log.ErrorContext(ctx, "Failed to get user", "err", err)
			return
		}
	}
	JSONResponse(w, http.StatusOK, user)
}

// GetUsers Get list of users
// @Summary Get list of users
// @Description Get list of all users from DB
// @Tags users
// @Accept json
// @Produce json
// @Param limit query int false "Limit records per page (default 10)"
// @Param page query int false "Page number (default 1)"
// @Param order query string false "Sort order: asc or desc (default desc)"
// @Success 200 {object} models.ListOfUsersResponse
// @Failure 500 {object} handler.JSONError "Failed to get users"
// @Router /users [get]
func (h *Handler) GetUsers(w http.ResponseWriter, r *http.Request) {

	var (
		limit uint64 = 10
		page  uint64 = 1
	)

	if s := r.URL.Query().Get("limit"); s != "" {
		l, err := strconv.ParseUint(s, 10, 64)
		if err != nil {
			WriteError(w, http.StatusBadRequest, "Invalid limit")
			return
		}
		limit = l
	}

	if s := r.URL.Query().Get("page"); s != "" {
		p, err := strconv.ParseUint(s, 10, 64)
		if err != nil {
			WriteError(w, http.StatusBadRequest, "Invalid page")
			return
		}
		page = p
	}

	order := r.URL.Query().Get("order")
	if order != "asc" {
		order = "desc"
	}

	ctx := r.Context()

	users, err := h.userService.GetUsers(ctx, limit, page, order)
	if err != nil {
		WriteError(w, http.StatusInternalServerError, "Failed to get users")
		slogger.Log.ErrorContext(ctx, "Failed to get users", "err", err)
		return
	}
	JSONResponse(w, http.StatusOK, users)
}

// Delete User delete by ID
// @Summary Delete user by ID
// @Description Delete user by ID from DB
// @Tags users
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path string true "User UUID" format(uuid)
// @Success 200 {object} nil
// @Failure 400 {object} handler.JSONError "Invalid user ID"
// @Failure 403 {object} handler.JSONError "You can delete only your own account"
// @Failure 500 {object} handler.JSONError "Failed to delete user"
// @Router /users/{id} [delete]
func (h *Handler) Delete(w http.ResponseWriter, r *http.Request) {
	targetID, err := uuid.Parse(chi.URLParam(r, "id"))
	if err != nil {
		WriteError(w, http.StatusBadRequest, "Invalid user ID")
		return
	}
	target, err := h.userService.GetUserByID(r.Context(), targetID)
	if err != nil {
		WriteError(w, http.StatusInternalServerError, "Failed to get user")
		slogger.Log.ErrorContext(r.Context(), "Failed to get user", "err", err)
		return
	}
	user, err := middleware.GetUserFromContext(r.Context())
	if err != nil {
		WriteError(w, http.StatusUnauthorized, "unauthorized")
		slogger.Log.ErrorContext(r.Context(), "unauthorized", "error", err)
		return
	}

	ctx := r.Context()

	if !models.PermissionCheck(user, target) {
		WriteError(w, http.StatusForbidden, models.ErrPermissionDenied.Error())
		return
	}
	err = h.userService.Delete(ctx, targetID)
	if err != nil {
		switch {
		case errors.Is(err, models.ErrUserNotFound):
			WriteError(w, http.StatusNotFound, "User not found")
			return
		case errors.Is(err, models.ErrPermissionDenied):
			WriteError(w, http.StatusForbidden, models.ErrPermissionDenied.Error())
			return

		}
		WriteError(w, http.StatusInternalServerError, "Failed to delete user")
		slogger.Log.ErrorContext(ctx, "Failed to delete user", "err", err)
		return
	}

	JSONResponse(w, http.StatusOK, nil)
}

// Update User update by ID
// @Summary Update user by ID
// @Description Update user information by its UUID
// @Tags users
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path string true "User UUID" format(uuid)
// @Param input body models.UpdateUserRequest true "User update info"
// @Success 200 {object} models.UserResponse "Successfully updated"
// @Failure 400 {object} handler.JSONError "Invalid user ID"
// @Failure 403 {object} handler.JSONError "Permission denied"
// @Failure 500 {object} handler.JSONError "Failed to update user"
// @Router /users/{id} [put]
func (h *Handler) Update(w http.ResponseWriter, r *http.Request) {

	targetID, err := uuid.Parse(chi.URLParam(r, "id"))
	if err != nil {
		WriteError(w, http.StatusBadRequest, "Invalid user ID")
		return
	}

	user, err := middleware.GetUserFromContext(r.Context())
	if err != nil {
		WriteError(w, http.StatusUnauthorized, "unauthorized")
		slogger.Log.ErrorContext(r.Context(), "unauthorized", "error", err)
		return
	}
	slogger.Log.DebugContext(r.Context(), "Updating user", "targetID", targetID, "requester", user)

	target, err := h.userService.GetUserByID(r.Context(), targetID)
	if err != nil {
		WriteError(w, http.StatusInternalServerError, "Failed to get user")
		slogger.Log.ErrorContext(r.Context(), "Failed to get user", "err", err)
		return
	}

	if !models.PermissionCheck(user, target) {
		WriteError(w, http.StatusForbidden, models.ErrPermissionDenied.Error())
		return
	}

	var req models.UpdateUserRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		WriteError(w, http.StatusBadRequest, "Invalid request body")
		slogger.Log.ErrorContext(r.Context(), "Invalid request body", "err", err)
		return
	}
	ctx := r.Context()

	if req.Role != nil && user.Role != models.RoleAdmin {
		WriteError(w, http.StatusBadRequest, models.ErrPermissionDenied.Error())
		return
	}
	if req.Password != nil {
		if err := utils.ValidatePassword(*req.Password); err != nil {
			WriteError(w, http.StatusBadRequest, err.Error())
			return
		}
	}

	updatedUser, err := h.userService.Update(ctx, targetID, req)
	if err != nil {
		switch {
		case errors.Is(err, models.ErrUserNotFound):
			WriteError(w, http.StatusNotFound, "User not found")
			return
		}
		WriteError(w, http.StatusInternalServerError, "Failed to update user")
		slogger.Log.ErrorContext(ctx, "Failed to update user", "err", err)
		return
	}

	JSONResponse(w, http.StatusOK, updatedUser)
}

func (h *Handler) Translate(w http.ResponseWriter, r *http.Request) {
	slogger.Log.DebugContext(r.Context(), "Translate")
	var req models.TranslateReq
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		WriteError(w, http.StatusInternalServerError, "invalid request body")
		return
	}
	if req.SourceWord == "" && req.TargetWord == "" {
		WriteError(w, http.StatusInternalServerError, "bad request: both source and target words are empty")
		return
	}

	ctx := r.Context()

	user, err := middleware.GetUserFromContext(ctx)
	if err != nil {
		WriteError(w, http.StatusUnauthorized, "unauthorized")
		slogger.Log.ErrorContext(ctx, "unauthorized", "error", err)
		return
	}
	req.ID = user.ID
	slogger.Log.Debug("translate req %+v", "req", req)
	resp, err := h.wordsService.Translate(ctx, req)
	if err != nil {
		WriteError(w, http.StatusInternalServerError, "error translating")
		return
	}
	slogger.Log.Debug("translate resp %+v", "resp", resp)

	JSONResponse(w, http.StatusOK, resp)
}

func (h *Handler) NewWord(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	var req models.CreateReq

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		WriteError(w, http.StatusBadRequest, "invalid request body")
		slogger.Log.ErrorContext(ctx, "invalid request body", "error", err)
		return
	}
	slogger.Log.DebugContext(ctx, "Handler NewWord called with req", "req", req)

	user, err := middleware.GetUserFromContext(ctx)
	if err != nil {
		WriteError(w, http.StatusUnauthorized, "unauthorized")
		slogger.Log.ErrorContext(ctx, "unauthorized", "error", err)
		return
	}

	resp, err := h.wordsService.Create(ctx, req, user.ID)
	if err != nil {
		switch {
		case errors.Is(err, models.ErrWordAlreadyExists):
			WriteError(w, http.StatusConflict, "word already exists")
			slogger.Log.ErrorContext(ctx, "word already exists", "error", err)
			return
		case errors.Is(err, models.ErrDBTimeout):
			WriteError(w, http.StatusRequestTimeout, "timeout")
			slogger.Log.ErrorContext(ctx, "timeout", "error", err)
			return
		default:
			WriteError(w, http.StatusInternalServerError, "internal server error")
			slogger.Log.ErrorContext(ctx, "internal server error", "error", err)
			return
		}
	}

	slogger.Log.DebugContext(ctx, "Handler NewWord called with resp", "resp", resp)
	JSONResponse(w, http.StatusOK, resp)
}

func (h *Handler) StartLesson(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	user, err := middleware.GetUserFromContext(ctx)
	if err != nil {
		WriteError(w, http.StatusUnauthorized, "unauthorized")
		slogger.Log.ErrorContext(ctx, "unauthorized", "error", err)
		return
	}

	word, err := h.wordsService.LessonStart(ctx, user.ID)
	if err != nil {
		switch {
		case errors.Is(err, models.ErrNoWordsForLesson):
			WriteError(w, http.StatusNotFound, "no words found for lesson")
			slogger.Log.ErrorContext(ctx, "no words found for lesson", "error", err)
			return
		default:
			WriteError(w, http.StatusInternalServerError, "internal server error")
			slogger.Log.ErrorContext(ctx, "internal server error", "error", err)
			return
		}
	}

	JSONResponse(w, http.StatusOK, word)

}

func (h *Handler) CheckAnswer(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	user, err := middleware.GetUserFromContext(ctx)
	if err != nil {
		WriteError(w, http.StatusUnauthorized, "unauthorized")
		slogger.Log.ErrorContext(ctx, "unauthorized", "error", err)
		return
	}

	var req models.AnswerReq
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		WriteError(w, http.StatusBadRequest, "invalid request body")
		slogger.Log.ErrorContext(ctx, "invalid request body", "error", err)
		return
	}
	slogger.Log.DebugContext(ctx, "Handler CheckAnswer called with req", "req", req)

	if req.ID == "" {
		slogger.Log.ErrorContext(ctx, "Handler CheckAnswer called with empty req.ID", "req", req.ID)
		WriteError(w, http.StatusBadRequest, "invalid request id")
		return
	}
	if req.TargetWord == "" {
		slogger.Log.ErrorContext(ctx, "Handler CheckAnswer called with empty req.TargetWord", "req", req.TargetWord)
		WriteError(w, http.StatusBadRequest, "invalid request target word")
		return
	}
	isCorrect, resp, err := h.wordsService.CheckAnswer(ctx, req, user.ID)
	if err != nil {
		WriteError(w, http.StatusInternalServerError, err.Error())
		return
	}
	responseData := struct {
		IsCorrect bool             `json:"is_correct"`
		NextWord  *models.Response `json:"next_word"`
	}{
		IsCorrect: isCorrect,
		NextWord:  resp,
	}

	if isCorrect {
		slogger.Log.DebugContext(ctx, "Handler CheckAnswer is correct")
	} else {
		slogger.Log.DebugContext(ctx, "Handler CheckAnswer is NOT correct")
	}

	JSONResponse(w, http.StatusOK, responseData)

}

func (h *Handler) Finish(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	user, err := middleware.GetUserFromContext(ctx)
	if err != nil {
		WriteError(w, http.StatusUnauthorized, "unauthorized")
		slogger.Log.ErrorContext(ctx, "unauthorized", "error", err)
		return
	}

	h.wordsService.Finish(ctx, user.ID)
	JSONResponse(w, http.StatusOK, nil)
}